requirements/base.txt
django==1.8.7
django-contrib-comments==1.6.1
ipython
django-extensions==1.5.9
djangorestframework==2.3.10
facebook-sdk==1.0.0
psycopg2==2.5.3
python-mimeparse==0.1.4
six==1.10.0
wsgiref==0.1.2
django-staticfiles-fontawesome==2.0.0
django-endless-pagination==2.0
pyreadline==2.0
pytz==2015.7
requests==2.10.0
unidecode==0.04.14
dropbox==2.0.0
pandas==0.17.1
openpyxl==2.1.3
python-dateutil==2.4.2
django-tastypie==0.12.2
pysolr
django-haystack
futures
django-blog-zinnia==0.15.2
django-polymorphic==0.7.2
jsonfield==1.0.0
autotag
django-autotag
django-taggit
unicode-slugify
solid_i18n==1.1.1
dj-database-url==0.4.1
networkx==1.11
langdetect==1.0.6
functools32==3.2.3-2
django-waffle==0.11.1

django-debug-toolbar>=1.0
pillow
feedparser
django-mailer
django-activity-stream==0.6.0
django-pagination==1.0.7
django-tagging==0.3.6
django-modeltranslation==0.10.2
django-cors-headers==1.1.0
factory_boy==2.9.2